# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Church ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /songs:
    get:
      operationId: listSongs
      tags: [songs]
      summary: List songs
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer } }
        - { in: query, name: limit, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongsListResponse"
    post:
      operationId: createSong
      tags: [songs]
      summary: Create a song
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSongRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
  /songs/suggestions:
    get:
      operationId: listSongSuggestions
      tags: [songs]
      summary: List song suggestions
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongSuggestionsListResponse"
    post:
      operationId: createSongSuggestion
      tags: [songs]
      summary: Create a song suggestion
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSongSuggestionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongSuggestion"
  /songs/{id}:
    get:
      operationId: getSongDetail
      tags: [songs]
      summary: Get song detail
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    put:
      operationId: updateSong
      tags: [songs]
      summary: Update a song
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSongRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    delete:
      operationId: deleteSong
      tags: [songs]
      summary: Delete a song
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /songs/suggestions/{id}:
    put:
      operationId: reviewSongSuggestion
      tags: [songs]
      summary: Review a song suggestion
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewSongSuggestionRequest"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongSuggestion"

  # ─── CULTOS ─────────────────────────────────────────────────────────────
...
  /cultos/{id}/songs:
    post:
      operationId: addCultoSong
      tags: [cultos]
      summary: Add song to culto
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCultoSongRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CultoSong"
...
  /cultos/{id}/songs/{songEntryId}:
    put:
      operationId: updateCultoSong
      tags: [cultos]
      summary: Update song notes
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: songEntryId, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCultoSongRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CultoSong"
    delete:
      operationId: deleteCultoSong
      tags: [cultos]
      summary: Remove song from culto
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: songEntryId, required: true, schema: { type: string } }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── COUNCIL ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Song:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        author: { type: string, nullable: true }
        songKey: { type: string, nullable: true }
        tempo: { type: string, nullable: true }
        lyrics: { type: string, nullable: true }
        chordChart: { type: string, nullable: true }
        category: { type: string, enum: [louvor, adoracao, hino, especial] }
        youtubeUrl: { type: string, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, title, category, createdAt, updatedAt]
    SongsListResponse:
      type: object
      properties:
        songs:
          type: array
          items:
            $ref: "#/components/schemas/Song"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [songs, total]
    CreateSongRequest:
      type: object
      properties:
        title: { type: string }
        youtubeUrl: { type: string }
      required: [title, youtubeUrl]
    UpdateSongRequest:
      type: object
      properties:
        title: { type: string }
        youtubeUrl: { type: string }
    SongSuggestion:
      type: object
      properties:
        id: { type: string }
        songId: { type: string, nullable: true }
        title: { type: string }
        url: { type: string }
        reason: { type: string, nullable: true }
        status: { type: string, enum: [pendente, aprovada, rejeitada] }
        reviewNote: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, title, url, status, createdAt]
    SongSuggestionsListResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/SongSuggestion"
        total: { type: integer }
      required: [suggestions, total]
    CreateSongSuggestionRequest:
      type: object
      properties:
        title: { type: string }
        url: { type: string }
        reason: { type: string }
      required: [title, url, reason]
    ReviewSongSuggestionRequest:
      type: object
      properties:
        status: { type: string, enum: [aprovada, rejeitada] }
        reviewNote: { type: string }
      required: [status]

    # ─── CULTOS ─────────────────────────────────────────────────────────────
...
    CultoSong:
      type: object
      properties:
        id: { type: string }
        cultoId: { type: string }
        songId: { type: string }
        songTitle: { type: string }
        order: { type: integer }
        notes: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, cultoId, songId, songTitle, order]
...
    AddCultoSongRequest:
      type: object
      properties:
        songId: { type: string }
        notes: { type: string }
      required: [songId]
    UpdateCultoSongRequest:
      type: object
      properties:
        notes: { type: string, nullable: true }
...